requirements.txt
faiss-cpu
sentence_transformers
streamlit
PyPDF2
flask